pyproject.toml
[project]
name = "crondashboard"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "croniter>=6.0.0",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "python-crontab>=3.3.0",
]
